+++ internal/xsql/xtable/tx_fake.go
@@ -7,7 +7,6 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/params"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/common"
 )
 
@@ -47,11 +46,11 @@ func beginTxFake(ctx context.Context, c *Conn) common.Tx {
 
 func (t *txFake) Commit(ctx context.Context) (err error) {
 	if !t.conn.isReady() {
-		return badconn.Map(xerrors.WithStackTrace(xerrors.Retryable(errNotReadyConn,
+		return xerrors.WithStackTrace(xerrors.Retryable(errNotReadyConn,
 			xerrors.Invalid(t),
 			xerrors.Invalid(t.conn),
 			xerrors.Invalid(t.conn.session),
-		)))
+		))
 	}
 
 	return nil
@@ -59,11 +58,11 @@ func (t *txFake) Commit(ctx context.Context) (err error) {
 
 func (t *txFake) Rollback(ctx context.Context) (err error) {
 	if !t.conn.isReady() {
-		return badconn.Map(xerrors.WithStackTrace(xerrors.Retryable(errNotReadyConn,
+		return xerrors.WithStackTrace(xerrors.Retryable(errNotReadyConn,
 			xerrors.Invalid(t),
 			xerrors.Invalid(t.conn),
 			xerrors.Invalid(t.conn.session),
-		)))
+		))
 	}
 
 	return err